src/integrations/slack/requirements.txt
slack-sdk>=3.21.3
python-dotenv>=1.0.0
typing-extensions>=4.5.0 